[Bug 134] Drop the "environment" concept.
Instead, allow the user to specify a Tddium server to connect to on the
CLI or with environment variables.

Add insecure mode to conditionally allow connecting to a server with an untrusted cert.

Dup the error message to get around this issue in highline (fix not yet
published): https://github.com/JEG2/highline/pull/81